+++ CHANGELOG.md
@@ -11,10 +11,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 ## [1.2.0] - 2020-01-04
+### Added
+- Add stop method.
+- Add reactivity to cli option changes. Cli is automatically started or stopped when cli option changes.
+
 ### Changed
 - Upgrade "@mocks-server" core dependency to v1.3.0.
 - Upgrade devDependencies.
 - Remove usage of core deprecated methods.
+
+### Fixed
+- Do not display behaviors inquirer if there are no behaviors to select.
 
 ## [1.1.0] - 2019-12-07
 ### Changed